+++ src/main/java/com/robertx22/mine_and_slash/database/spells/entities/cloud/RockSlideEntity.java
@@ -57,8 +57,7 @@ public class RockSlideEntity extends BaseCloudEntity {
             EntitySpellData sdata = getSpellData();
             int tickrate = sdata.configs.get(SC.TICK_RATE)
                     .intValue();
-            int RADIUS = sdata.configs.get(SC.RADIUS)
-                    .intValue();
+            float RADIUS = sdata.configs.get(SC.RADIUS);
 
             super.onTick();
