+++ src/main/java/com/robertx22/mine_and_slash/database/spells/entities/proj/BlastTrapEntity.java
@@ -58,8 +58,7 @@ public class BlastTrapEntity extends EntityBaseProjectile {
     public void onTick() {
 
         EntitySpellData sdata = getSpellData();
-        int RADIUS = sdata.configs.get(SC.RADIUS)
-            .intValue();
+        float RADIUS = sdata.configs.get(SC.RADIUS);
 
         if (this.inGround || this.ticksExisted % 5 == 0) {
             if (!world.isRemote) {
